Move utils from libs folder to it's own
